refactor datalake so that csv parsing is modularized
