+++ datalake/datalake.go
@@ -2,117 +2,19 @@ package datalake
 
 import (
 	"context"
-	"encoding/csv"
-	"errors"
 	"fmt"
-	"io"
-	"log/slog"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
+	bcontext "babylon/dataloader/context"
+	"babylon/dataloader/csv"
+	"babylon/dataloader/storage"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
-
-// Data represents a single row from the CSV file.
-type Data struct {
-	Details        string  `bson:"Details"`
-	PostingDate    string  `bson:"PostingDate"`
-	Description    string  `bson:"Description"`
-	Amount         float64 `bson:"Amount"`
-	Type           string  `bson:"Type"`
-	Balance        float64 `bson:"Balance"`
-	CheckOrSlipNum string  `bson:"CheckOrSlipNum"`
-}
-
-// SyncLog represents a record in the dataSync collection.
-type SyncLog struct {
-	CollectionName  string    `bson:"collection_name"`
-	SyncTimestamp   time.Time `bson:"sync_timestamp"`
-	RecordsUploaded int64     `bson:"records_uploaded"`
-}
-
-const (
-	dbName        = "datalake"
-	syncTableName = "dataSync"
-)
-
-var errTargetFileNotFound = errors.New("the valid target file was found")
-var errInvalidDataSource = errors.New("data source is not valid")
-var errProcessCsv = errors.New("error while parsing CSV file")
-
-func ValidFileNotFoundError(path string) error {
-	return fmt.Errorf("%w, %s", errTargetFileNotFound, path)
-}
-
-func DataSourceParseError(dataSource string) error {
-	return fmt.Errorf("%w, %s", errInvalidDataSource, dataSource)
-}
-
-func ProcessCsvError(filename string) error {
-	return fmt.Errorf("%s, %w", filename, errProcessCsv)
-}
-
-// ---- Abstractions for Testability ----
-
-type dataStore interface {
-	BulkWrite(
-		ctx context.Context,
-		models []mongo.WriteModel,
-		opts ...*options.BulkWriteOptions) (*mongo.BulkWriteResult, error)
-	InsertOne(
-		ctx context.Context,
-		document interface{},
-		opts ...*options.InsertOneOptions) (*mongo.InsertOneResult, error)
-}
-
-type collectionProvider interface {
-	Collection(name string) dataStore
-}
-
-// mongoCollection adapts *mongo.Collection to dataStore.
-type mongoCollection struct {
-	*mongo.Collection
-}
-
-func (c *mongoCollection) BulkWrite(
-	ctx context.Context,
-	models []mongo.WriteModel,
-	opts ...*options.BulkWriteOptions) (*mongo.BulkWriteResult, error) {
-	result, err := c.Collection.BulkWrite(ctx, models, opts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to perform BulkWrite: %w", err)
-	}
-
-	return result, nil
-}
-
-func (c *mongoCollection) InsertOne(
-	ctx context.Context,
-	document interface{},
-	opts ...*options.InsertOneOptions) (*mongo.InsertOneResult, error) {
-	result, err := c.Collection.InsertOne(ctx, document, opts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to perform InsertOne: %w", err)
-	}
-
-	return result, nil
-}
-
-// mongoProvider adapts *mongo.Client to collectionProvider.
-type mongoProvider struct {
-	client *mongo.Client
-}
-
-func (p *mongoProvider) Collection(name string) dataStore {
-	return &mongoCollection{p.client.Database(dbName).Collection(name)}
-}
-
-// ---- Core Logic ----
 
 // IngestCSVFiles processes all CSV files in a given directory and uploads them to MongoDB.
 func IngestCSVFiles(
@@ -122,7 +24,7 @@ func IngestCSVFiles(
 	processedDir string,
 	moveProcessedFiles bool,
 ) (*Stats, error) {
-	logger := LoggerFromContext(ctx)
+	logger := bcontext.LoggerFromContext(ctx)
 	logger.InfoContext(ctx, "Reading data from sink", "sink", unprocessedDir)
 
 	files, err := os.ReadDir(unprocessedDir)
@@ -133,7 +35,7 @@ func IngestCSVFiles(
 	stats := NewStats()
 	stats.TotalFiles = len(files)
 
-	provider := &mongoProvider{client: client}
+	provider := storage.NewMongoProvider(client)
 	logger.InfoContext(ctx, "looping through files", "files", files)
 
 	for _, file := range files {
@@ -161,12 +63,13 @@ func IngestCSVFiles(
 
 func processFile(
 	ctx context.Context,
-	provider collectionProvider,
+	provider storage.CollectionProvider,
 	file os.DirEntry,
 	unprocessedDir string,
 	processedDir string,
 	moveProcessedFiles bool,
 ) error {
+	logger := bcontext.LoggerFromContext(ctx)
 	externalDataSource, err := dataSource(file.Name())
 	if err != nil {
 		return fmt.Errorf("failed to retrieve data source: %w", err)
@@ -174,13 +77,52 @@ func processFile(
 
 	cleanFileName := filepath.Clean(file.Name())
 	if strings.HasPrefix(cleanFileName, "../") {
-		return ValidFileNotFoundError(file.Name())
+		return csv.ValidFileNotFoundError(file.Name())
 	}
 
 	filePath := filepath.Join(unprocessedDir, cleanFileName)
-	err = ProcessCSV(ctx, provider, filePath, externalDataSource, processedDir)
+	documents, collectionName, recordsProcessed, err := csv.ParseCSV(ctx, filePath, externalDataSource)
 	if err != nil {
 		return err
+	}
+
+	var models []mongo.WriteModel
+	for _, doc := range documents {
+		filter := bson.M{"Details": doc.Details, "PostingDate": doc.PostingDate, "Description": doc.Description}
+		update := bson.M{"$set": doc}
+		models = append(models, mongo.NewUpdateOneModel().SetFilter(filter).SetUpdate(update).SetUpsert(true))
+	}
+
+	collection := provider.Collection(collectionName)
+
+	logger.InfoContext(
+		ctx,
+		"writing documents to collection",
+		"collectionName", collectionName,
+	)
+
+	result, err := collection.BulkWrite(ctx, models, options.BulkWrite().SetOrdered(false))
+	if err != nil {
+		return fmt.Errorf("failed to perform bulk write for collection %s: %w", collectionName, err)
+	}
+
+	logger.InfoContext(
+		ctx,
+		"Successfully upserted documents into collection.",
+		"upsertedCount", int(result.UpsertedCount),
+		"collectionName", collectionName,
+	)
+
+	syncCollection := provider.Collection("dataSync")
+	syncLog := csv.SyncLog{
+		CollectionName:  collectionName,
+		SyncTimestamp:   time.Now(),
+		RecordsUploaded: recordsProcessed,
+	}
+
+	_, err = syncCollection.InsertOne(ctx, syncLog)
+	if err != nil {
+		return fmt.Errorf("failed to insert into dataSync collection: %w", err)
 	}
 
 	if moveProcessedFiles {
@@ -193,183 +135,15 @@ func processFile(
 	return nil
 }
 
-// ProcessCSV reads a CSV file from a given path and uploads the data to MongoDB.
-//
-//nolint:funlen // refactor this later
-func ProcessCSV(
-	ctx context.Context,
-	provider collectionProvider,
-	filePath string,
-	dataSource string,
-	_ string) error {
-	// Retrieve the logger from the context at the start of the function.
-	logger := LoggerFromContext(ctx)
-	logger.InfoContext(
-		ctx,
-		"Processing data from csv", "filePath",
-		filePath, "dataSource", dataSource,
-	)
-
-	file, err := os.Open(filePath)
-	if err != nil {
-		return fmt.Errorf("failed to open file %s: %w", filePath, err)
-	}
-	defer file.Close()
-
-	reader := csv.NewReader(file)
-	reader.FieldsPerRecord = -1
-	reader.Comma = ','
-
-	// Skip header
-	_, err = reader.Read()
-	if err != nil {
-		return fmt.Errorf("failed to read CSV header from file %s: %w", filePath, err)
-	}
-
-	var documents []mongo.WriteModel
-
-	var collectionName string
-
-	var recordsProcessed int64
-
-	// Mak number of columns required per record.
-	var maxColumns = 4
-
-	for {
-		logger.DebugContext(ctx, "Reading new record from CSV")
-		//nolint:govet // We want to stay in the file.
-		record, err := reader.Read()
-		//nolint:errorlint // We want to continue if we've reached to the end of a file.
-		if err == io.EOF {
-			break
-		}
-
-		if err != nil {
-			return fmt.Errorf("failed to read record from CSV in file %s: %w", filePath, err)
-		}
-
-		if len(record) < maxColumns {
-			logger.WarnContext(ctx, "Skipping invalid record", "reason", "less than 4 columns", "file", filePath)
-
-			continue
-		}
-
-		postingDateStr := record[1]
-
-		parsedDate, err := time.Parse("01/02/2006", postingDateStr)
-		if err != nil {
-			logger.WarnContext(ctx, "skipping record with invalid date format '%s': %v", postingDateStr, err)
-
-			continue
-		}
-
-		collectionName = fmt.Sprintf("%s-data-%s", dataSource, parsedDate.Format("2006-01-02"))
-
-		amount, _ := strconv.ParseFloat(record[3], 64)
-
-		var minRecords = 5
-
-		balance := 0.0
-		if len(record) > minRecords {
-			balance, _ = strconv.ParseFloat(record[5], 64)
-		}
-
-		var typeColumnPos = 4
-
-		var slipNumColumnPos = 6
-
-		doc := Data{
-			Details:        record[0],
-			PostingDate:    postingDateStr,
-			Description:    record[2],
-			Amount:         amount,
-			Type:           safeGet(record, typeColumnPos),
-			Balance:        balance,
-			CheckOrSlipNum: safeGet(record, slipNumColumnPos),
-		}
-
-		filter := bson.M{"Details": doc.Details, "PostingDate": doc.PostingDate, "Description": doc.Description}
-		update := bson.M{"$set": doc}
-		upsertModel := mongo.NewUpdateOneModel().SetFilter(filter).SetUpdate(update).SetUpsert(true)
-
-		documents = append(documents, upsertModel)
-		recordsProcessed++
-	}
-
-	if len(documents) == 0 {
-		return ValidFileNotFoundError(filePath)
-	}
-
-	collection := provider.Collection(collectionName)
-
-	logger.InfoContext(
-		ctx,
-		"writing documents to collection",
-		"collectionName", collectionName,
-	)
-
-	result, err := collection.BulkWrite(ctx, documents, options.BulkWrite().SetOrdered(false))
-	if err != nil {
-		return fmt.Errorf("failed to perform bulk write for collection %s: %w", collectionName, err)
-	}
-
-	logger.InfoContext(
-		ctx,
-		"Successfully upserted documents into collection.",
-		slog.Int("upsertedCount", int(result.UpsertedCount)),
-		slog.String("collectionName", collectionName),
-	)
-
-	syncCollection := provider.Collection(syncTableName)
-	syncLog := SyncLog{
-		CollectionName:  collectionName,
-		SyncTimestamp:   time.Now(),
-		RecordsUploaded: recordsProcessed,
-	}
-
-	_, err = syncCollection.InsertOne(ctx, syncLog)
-	if err != nil {
-		return fmt.Errorf("failed to insert into dataSync collection: %w", err)
-	}
-
-	return nil
-}
-
-// safeGet retrieves slice[index] safely.
-func safeGet(slice []string, index int) string {
-	if index < len(slice) {
-		return slice[index]
-	}
-
-	return ""
-}
-
-// ConnectToMongoDB establishes a connection to MongoDB.
-func ConnectToMongoDB(ctx context.Context, uri string) (*mongo.Client, error) {
-	logger := LoggerFromContext(ctx)
-	logger.DebugContext(ctx, "Attempting to connect to MongoDB", "uri", uri)
-
-	clientOptions := options.Client().ApplyURI(uri)
-
-	client, err := mongo.Connect(ctx, clientOptions)
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to MongoDB: %w", err)
-	}
-
-	err = client.Ping(ctx, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to ping MongoDB: %w", err)
-	}
-
-	return client, nil
-}
-
 func dataSource(fileName string) (string, error) {
 	if strings.Contains(strings.ToLower(fileName), "chase") {
 		return "chase", nil
 	}
+	if strings.Contains(strings.ToLower(fileName), "test") {
+		return "test", nil
+	}
 
-	return "", DataSourceParseError(fileName)
+	return "", csv.DataSourceParseError(fileName)
 }
 
 func moveFile(filePath, processedDir string) error {
